requirements.txt
## made on python 3.14

fastapi
uvicorn[standard]
websockets
python-dotenv
numpy
requests
httpx
groq

# ai|coustics standalone SDK (local noise cancellation)
aic-sdk
